Migrate `utils` to GNU Make.
