+++ .clangd
@@ -1,2 +1,2 @@
 CompileFlags:
-	Add: [-std=c++20, -Wall, -pedantic-errors]
+	Add: [-std=c++20, -Wall, -pedantic-errors, -I../../lib, -I../../../lib]
