+++ pyrightconfig.json
@@ -1,4 +1,4 @@
 {
 	"pythonVersion": "3.8",
-	"extraPaths": ["utils/core-py"]
+	"extraPaths": ["utils/lib-py"]
 }
