--- utils/cbundle/.gitignore
@@ -1,2 +1,0 @@
-cbundle
-cbundle.exe
