--- utils/convsym/.gitignore
@@ -1,2 +1,0 @@
-ConvSym.exe
-convsym
