--- utils/convsym/build.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-set -e
-
-cc -std=c++20 -O3 -pedantic-errors -s -Wall convsym.cpp -o convsym -lstdc++
-cp convsym ../../modules/exec/convsym
