+++ utils/convsym/src/input/ASM68K_Listing.cpp
@@ -9,8 +9,8 @@
 #include <map>
 #include <set>
 
-#include "../../core/IO.hpp"
-#include "../../core/OptsParser.hpp"
+#include <IO.hpp>
+#include <OptsParser.hpp>
 
 #include "InputWrapper.hpp"
