+++ utils/convsym/src/input/AS_Listing.cpp
@@ -10,8 +10,8 @@
 #include <string_view>
 #include <map>
 
-#include "../../core/IO.hpp"
-#include "../../core/OptsParser.hpp"
+#include <IO.hpp>
+#include <OptsParser.hpp>
 
 #include "InputWrapper.hpp"
